+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - node@5.10.1, canvas@1.3.14, jsdom@8.5.0 support
 - `Konva.Path` will be filled when it is not closed
 - `Animation.start()` will not not immediate sync draw. This should improve performance a little.
+- Warning when node for `Tween` is not in layer yet.
 
 ## [0.12.4][2016-04-19]